backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
psycopg2-binary
openai
requests
qdrant-client
python-dotenv
